testing different bytes from decompiled SMB
I added a condition to make Mario jump if he encountered an obstacle, which failed. I also imported and set various values straight from the ROM onto the GUI of FCEUX. With those, I tested Mario's x and y positions and speeds with messy variable names just to see whether or not they would change and if I could use those for conditionals to train the bot.